Fix compile issues, initial ICorDebug work
Fix some compile issues, begin to implement ICorDebug support.
