+++ src/Microsoft.Diagnostics.Runtime/Desktop/helpers.cs
@@ -267,236 +267,4 @@ namespace Microsoft.Diagnostics.Runtime.Desktop
         }
         #endregion
     }
-
-    internal class NativeMethods
-    {
-        public static bool LoadNative(string dllName)
-        {
-            return LoadLibrary(dllName) != IntPtr.Zero;
-        }
-
-        private const string Kernel32LibraryName = "kernel32.dll";
-
-        public const uint FILE_MAP_READ = 4;
-
-        // Call CloseHandle to clean up.
-        [DllImport(Kernel32LibraryName, SetLastError = true)]
-        public static extern SafeWin32Handle CreateFileMapping(
-           SafeFileHandle hFile,
-           IntPtr lpFileMappingAttributes, PageProtection flProtect, uint dwMaximumSizeHigh,
-           uint dwMaximumSizeLow, string lpName);
-
-        [DllImport(Kernel32LibraryName, SetLastError = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool UnmapViewOfFile(IntPtr baseAddress);
-
-
-        [DllImport(Kernel32LibraryName, SetLastError = true)]
-        public static extern SafeMapViewHandle MapViewOfFile(SafeWin32Handle hFileMappingObject, uint
-           dwDesiredAccess, uint dwFileOffsetHigh, uint dwFileOffsetLow,
-           IntPtr dwNumberOfBytesToMap);
-
-        [DllImportAttribute(Kernel32LibraryName)]
-        public static extern void RtlMoveMemory(IntPtr destination, IntPtr source, IntPtr numberBytes);
-
-        [DllImport(Kernel32LibraryName, SetLastError = true, PreserveSig = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool CloseHandle(IntPtr handle);
-
-        [DllImportAttribute(Kernel32LibraryName)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        internal static extern bool FreeLibrary(IntPtr hModule);
-
-        public static IntPtr LoadLibrary(string lpFileName)
-        {
-            return LoadLibraryEx(lpFileName, 0, LoadLibraryFlags.NoFlags);
-        }
-
-        [DllImportAttribute(Kernel32LibraryName, SetLastError = true)]
-        public static extern IntPtr LoadLibraryEx(String fileName, int hFile, LoadLibraryFlags dwFlags);
-
-        [Flags]
-        public enum LoadLibraryFlags : uint
-        {
-            NoFlags = 0x00000000,
-            DontResolveDllReferences = 0x00000001,
-            LoadIgnoreCodeAuthzLevel = 0x00000010,
-            LoadLibraryAsDatafile = 0x00000002,
-            LoadLibraryAsDatafileExclusive = 0x00000040,
-            LoadLibraryAsImageResource = 0x00000020,
-            LoadWithAlteredSearchPath = 0x00000008
-        }
-
-
-        [Flags]
-        public enum PageProtection : uint
-        {
-            NoAccess = 0x01,
-            Readonly = 0x02,
-            ReadWrite = 0x04,
-            WriteCopy = 0x08,
-            Execute = 0x10,
-            ExecuteRead = 0x20,
-            ExecuteReadWrite = 0x40,
-            ExecuteWriteCopy = 0x80,
-            Guard = 0x100,
-            NoCache = 0x200,
-            WriteCombine = 0x400,
-        }
-
-        [DllImport("kernel32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        private static extern bool IsWow64Process([In] IntPtr hProcess, [Out] out bool isWow64);
-
-        [DllImport("version.dll")]
-        internal static extern bool GetFileVersionInfo(string sFileName, int handle, int size, byte[] infoBuffer);
-
-        [DllImport("version.dll")]
-        internal static extern int GetFileVersionInfoSize(string sFileName, out int handle);
-
-        [DllImport("version.dll")]
-        internal static extern bool VerQueryValue(byte[] pBlock, string pSubBlock, out IntPtr val, out int len);
-
-        private const int VS_FIXEDFILEINFO_size = 0x34;
-        public static short IMAGE_DIRECTORY_ENTRY_COM_DESCRIPTOR = 14;
-
-#if !V2_SUPPORT
-        [DefaultDllImportSearchPaths(DllImportSearchPath.LegacyBehavior)]
-#endif
-        [DllImport("dbgeng.dll")]
-        internal static extern uint DebugCreate(ref Guid InterfaceId, [MarshalAs(UnmanagedType.IUnknown)] out object Interface);
-        [UnmanagedFunctionPointer(CallingConvention.StdCall)]
-        internal delegate int CreateDacInstance([In, ComAliasName("REFIID")] ref Guid riid,
-                                       [In, MarshalAs(UnmanagedType.Interface)] IDacDataTarget data,
-                                       [Out, MarshalAs(UnmanagedType.IUnknown)] out object ppObj);
-
-
-        [DllImport("kernel32.dll")]
-        internal static extern IntPtr GetProcAddress(IntPtr hModule, string procedureName);
-
-        internal static bool IsEqualFileVersion(string file, VersionInfo version)
-        {
-            int major, minor, revision, patch;
-            if (!GetFileVersion(file, out major, out minor, out revision, out patch))
-                return false;
-
-            return major == version.Major && minor == version.Minor && revision == version.Revision && patch == version.Patch;
-        }
-
-
-        internal static bool GetFileVersion(string dll, out int major, out int minor, out int revision, out int patch)
-        {
-            major = minor = revision = patch = 0;
-
-            int handle;
-            int len = GetFileVersionInfoSize(dll, out handle);
-
-            if (len <= 0)
-                return false;
-
-            byte[] data = new byte[len];
-            if (!GetFileVersionInfo(dll, handle, len, data))
-                return false;
-
-            IntPtr ptr;
-            if (!VerQueryValue(data, "\\", out ptr, out len))
-            {
-                return false;
-            }
-
-
-            byte[] vsFixedInfo = new byte[len];
-            Marshal.Copy(ptr, vsFixedInfo, 0, len);
-
-            minor = (ushort)Marshal.ReadInt16(vsFixedInfo, 8);
-            major = (ushort)Marshal.ReadInt16(vsFixedInfo, 10);
-            patch = (ushort)Marshal.ReadInt16(vsFixedInfo, 12);
-            revision = (ushort)Marshal.ReadInt16(vsFixedInfo, 14);
-
-            return true;
-        }
-
-        internal static bool TryGetWow64(IntPtr proc, out bool result)
-        {
-            if (Environment.OSVersion.Version.Major > 5 ||
-                (Environment.OSVersion.Version.Major == 5 && Environment.OSVersion.Version.Minor >= 1))
-            {
-                return IsWow64Process(proc, out result);
-            }
-            else
-            {
-                result = false;
-                return false;
-            }
-        }
-    }
-
-
-    internal sealed class SafeWin32Handle : SafeHandleZeroOrMinusOneIsInvalid
-    {
-        public SafeWin32Handle() : base(true) { }
-
-        public SafeWin32Handle(IntPtr handle)
-            : this(handle, true)
-        {
-        }
-
-        public SafeWin32Handle(IntPtr handle, bool ownsHandle)
-            : base(ownsHandle)
-        {
-            SetHandle(handle);
-        }
-
-        protected override bool ReleaseHandle()
-        {
-            return NativeMethods.CloseHandle(handle);
-        }
-    }
-
-    internal sealed class SafeMapViewHandle : SafeHandleZeroOrMinusOneIsInvalid
-    {
-        private SafeMapViewHandle() : base(true) { }
-
-        protected override bool ReleaseHandle()
-        {
-            return NativeMethods.UnmapViewOfFile(handle);
-        }
-
-        // This is technically equivalent to DangerousGetHandle, but it's safer for file
-        // mappings. In file mappings, the "handle" is actually a base address that needs
-        // to be used in computations and RVAs.
-        // So provide a safer accessor method.
-        public IntPtr BaseAddress
-        {
-            get
-            {
-                return handle;
-            }
-        }
-    }
-
-    internal sealed class SafeLoadLibraryHandle : SafeHandleZeroOrMinusOneIsInvalid
-    {
-        private SafeLoadLibraryHandle() : base(true) { }
-        public SafeLoadLibraryHandle(IntPtr handle)
-            : base(true)
-        {
-            SetHandle(handle);
-        }
-
-        protected override bool ReleaseHandle()
-        {
-            return NativeMethods.FreeLibrary(handle);
-        }
-
-        // This is technically equivalent to DangerousGetHandle, but it's safer for loaded
-        // libraries where the HMODULE is also the base address the module is loaded at.
-        public IntPtr BaseAddress
-        {
-            get
-            {
-                return handle;
-            }
-        }
-    }
 }
