+++ src/Microsoft.Diagnostics.Runtime/v2/Microsoft.Diagnostics.Runtime.v2.csproj
@@ -92,6 +92,10 @@
     <Compile Include="..\Debugger\IDebugSystemObjects2.cs" />
     <Compile Include="..\Debugger\IDebugSystemObjects3.cs" />
     <Compile Include="..\Debugger\Structs.cs" />
+    <Compile Include="..\ICorDebug\Helpers.cs" />
+    <Compile Include="..\ICorDebug\ICorDebugWrappers.cs" />
+    <Compile Include="..\ICorDebug\IMetadataImport.cs" />
+    <Compile Include="..\ICorDebug\IMetaHostWrappers.cs" />
     <Compile Include="..\Desktop\com.cs" />
     <Compile Include="..\DacInterfaces.cs" />
     <Compile Include="..\Desktop\lockinspection.cs" />
